requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
cryptography==42.0.5
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
google-auth==2.29.0
firebase-admin==6.5.0
python-multipart==0.0.9
alembic==1.13.1
google-generativeai==0.5.4